Don't change directory before running tests
Don't change the working directory to civicrm/tools before running the
test.

It appears to be entirely unnecessary as far as I can tell -- while it
created multiple problems: it was not possible to use getcwd() within
the test classes (which might or might not be a good idea... not sure);
and more importatly, running multiple tests at once was more awkward, as
a simple "civix test tests/phpunit/..." wouldn't work. (Requiring an
absolute path instead.)