generate:api - Explicitly prefer singleword action names
The proximate reason for change -- `tests/make-example.sh` was using
CamelCase actions, and it passed on macOS, but it didn't pass on Linux.
I'd like the test to give consistent output on both platforms.

It may be that there are semi-legit ways to get actions working with
CamelCase and/or under_score_case, e.g.

* Use a file-name like `MyEntity/Myaction.php` but call it as `MyEntity, MyAction`.
  This seemed to work in one Linux trial for calling `MyAction`, but calls weren't
  interchangeable with `my_action`. It seemed like an unintuitive arrangement, and
  I haven't re-tested to be sure.
* Don't generate standalone action files. Put all the actions into
  a combined per-entity file. However, this doesn't fit well into civix's
  mental model of "call generator => produce well-defined PHP file".

In any case, I don't really want to futz with this much more since APIv4 is
preferred for new development.

This change just further enshrines the work-around we've been using - ie
don't try CamelCase or under_score_case, just use singleword.