Use CRM_Extension_Upgrader_Base and info.xml <upgrader> tag
Since 5.38 core supports an <upgrader> tag in liu of hooks, and a common base class.
For extensions where core version is >= 5.38, this will generate upgrader classes that
use the common base, and the tag instead of delegated hooks.
It will also perform this cleanup on existing modules during each `civix upgrade`.