Drop unused "codecept" integration
I never used this, and I can't find anyone in `universe` who has published
code with it.  It adds makework/dependencies/weight for QA of civix, and
there's no point.

(If someone did want to follow-through on an intitiative for codecept,
then...  I think that would be a de-facto new initiative.  Maybe this patch
would be a useful reference...  but that doesn't mean we need to maintain it
speculatively...)