upgrades/23.02.1 - Suggest enabling the "Civi" namespace
This came up on Mattermost as something that frequently trips people up.

For many authors with existing extensions, it's probably better to enable this proactively.
Many folks don't think about classloading, and (if they omit this) it can take an hour of
debugging to figure out why their classes don't load.

Still, this is not strongly required.  Many existing extensions work fine
without it.  There's a (small) performance argument for keeping it disabled
if you don't need it.

So this update presents the option.  Even if the developer say no, it should
at least promote some consciousness about classloading (*and thereby reduce
future debug times*).