When exporting, prevent conflicts between modules
For example:

* Install `civi_campaign` and `civicrm_admin_ui`
* Run `cv api4 SavedSearch.get +s id,name -T`. Note the id of `SavedSearch_Administer_Campaigns`
* In `civicrm_admin_ui`, export `SavedSearch_Administer_Campaigns`
* Problem: We now have two modules trying to provide `SavedSearch_Administer_Campaigns` (with the same value of `civicrm_saved_search.name`)

The mistake here is that this search is actually owned by `civi_campaign`.
Allowing `civicrm_admin_ui` to make a copy with the exact same content (but
a different `civicrm_managed` record) ensures a conflict. This situation is
most likely a user-error (e.g. you forgot where the mgd was supposed to
live, or you didn't pay attention to PWD).

`