Replace `CRM_Core_DAO_Base` with CiviMix\Schema\*\DAO
In this arrangement, the anonymous DAO implementation is used on any
environment.

This means that it will choose the highest available version of the DAO-base
at the moment of loading, so it presumptively supports future backporting.

However, there may be some micro-optimization questions in this implementation.

This variant has... OK support for IDE hints/PHPStorm. The drill-down will
technically point to the wrong file (eg taking you to
`CRM/Core/DAO/Base.php` instead of `mixin/lib/civimix-schema/src/DAO.php`).
However, the autocomplete suggestions should be on-point.