Update references from extension aliases to the CRM_Core_Extension
Prior to CiviCRM 5.52 we threw a mismash of these extensions. They all had to work everywhere because it
was unpredictable. In 5.52 we made the API_Exception and CiviCRM_Api3_Exception aliases of
CRM_Core_Exception. However, on the off chance people are creating extensions for really old civi
versions with the latest civix it would still be handled on those older versions (we were
always throwing CRM_Core_Exception all over the show)